pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cogman-guard"
version = "0.1.0"
description = "Operational analysis tools for embeddings & AI systems"
readme = "README.md"
authors = [{ name = "tiewphopum", email = "creator@cogmanframework.com" }]
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.21.0",
    "torch>=1.9.0",
    "matplotlib>=3.4.0",
    "scipy>=1.7.0",
    "scikit-learn>=0.24.0",
    "plotly>=5.0.0",
    "networkx>=2.6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
providers = [
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "anthropic>=0.18.0",
]

tools = [
    "build",
    "twine",
]

[project.scripts]
cogman-guard-quick-test = "cogman_tools.quick_cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
